requirements.txt
Flask==2.0.3
numpy==1.24.3
tensorflow==2.13.1
Werkzeug==2.2.2
pillow 